requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==2.1.1
openpyxl==3.1.5
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
rapidfuzz==3.10.0
streamlit==1.38.0
requests==2.32.3
xlsxwriter==3.2.0
pytest==8.3.3
httpx==0.27.2
loguru==0.7.2